requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
bcrypt
